initial_V4

+++ src/BmiService.java
@@ -1,5 +1,5 @@
 public class BmiService {
-    public double calculate(double heightInMeters, double weightInKilograms) {
+    public int calculate(double heightInMeters, double weightInKilograms) {
         double bmi = weightInKilograms / (heightInMeters * heightInMeters);
         return (int) bmi;
     }
